move compass app ops into separate files
